requirements.txt
click==8.1.7
dill==0.3.7
latexify-py==0.4.2
numpy==1.26.3
